Code style improvement. No unnecessary type conversion.

+++ Assets/SEE/Tools/FaceCam/FaceCam.cs
@@ -507,7 +507,7 @@ namespace SEE.Tools.FaceCam
                     // Apply the cutout texture size to the FacCam prefab.
                     // The size is way to big, so it needs to be reduced. A maximum height is used.
                     float divisor = croppedTextureHeight / MaxHeight;
-                    transform.localScale = new Vector3(((float)croppedTextureWidth) / divisor, (float)croppedTextureHeight / divisor, -1); // Without '(float)' the result is just '0'.
+                    transform.localScale = new Vector3((croppedTextureWidth) / divisor, croppedTextureHeight / divisor, -1);
                 }
 
                 // Copy the pixels from the original texture to the cutout texture.
